extract function: `addToSelection(map, params)`
